@@ -3,36 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.2",
+    "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@reduxjs/toolkit": "^1.6.1",
+    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@reduxjs/toolkit": "^1.7.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/d3-collection": "^1.0.10",
     "@types/d3-force": "^3.0.3",
-    "@types/react-redux": "^7.1.16",
+    "@types/react-redux": "^7.1.21",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
-    "@use-gesture/react": "^10.1.0",
-    "aws-amplify": "^4.3.2",
+    "@types/tinycolor2": "^1.4.3",
+    "@use-gesture/react": "^10.2.4",
+    "aws-amplify": "^4.3.11",
     "aws-amplify-react": "^3.1.8",
     "d3": "^5.16.0",
-    "dexie": "^3.0.3",
-    "immer": "^9.0.6",
+    "dexie": "^3.2.0",
+    "immer": "^9.0.7",
     "ohm-js": "^16.0.0",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.1",
-    "react-spring": "^9.0.0-rc.3",
-    "react-virtualized-auto-sizer": "^1.0.3",
-    "react-virtuoso": "^1.1.3",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^4.0.3",
+    "react-spring": "^9.3.3",
+    "react-virtualized-auto-sizer": "^1.0.6",
+    "react-virtuoso": "^1.11.1",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.1",
     "tinycolor2": "^1.4.2",
-    "typescript": "^4.2.3",
+    "typescript": "^4.5.4",
     "uuid": "^8.3.2"
   },
   "scripts": {
@@ -60,17 +61,17 @@
     ]
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^6.1.11",
+    "@storybook/addon-actions": "^6.4.9",
     "@storybook/addon-essentials": "^6.4.9",
-    "@storybook/addon-links": "^6.1.11",
-    "@storybook/node-logger": "^6.1.11",
-    "@storybook/preset-create-react-app": "^3.1.5",
+    "@storybook/addon-links": "^6.4.9",
+    "@storybook/node-logger": "^6.4.9",
+    "@storybook/preset-create-react-app": "^3.2.0",
     "@storybook/react": "^6.4.9",
-    "@types/jest": "^26.0.20",
-    "@types/react": "^17.0.19",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/uuid": "^8.3.0",
-    "lorem-ipsum": "^2.0.3",
+    "@types/jest": "^26.0.24",
+    "@types/react": "^17.0.38",
+    "@types/react-router-dom": "^5.3.2",
+    "@types/uuid": "^8.3.3",
+    "lorem-ipsum": "^2.0.4",
     "redux-mock-store": "^1.5.4"
   }
 }
