@@ -1,11 +1,7 @@
 import React, { FunctionComponent } from 'react'
 
 import {
-    Switch,
-    Route,
-    Link,
-    useRouteMatch,
-    useHistory
+    useNavigate
 } from "react-router-dom"
 
 import Box from '@material-ui/core/Box'
@@ -20,7 +16,6 @@ import Paper from '@material-ui/core/Paper'
 import { makeStyles } from "@material-ui/core/styles"
 
 import useStyles from '../../styles'
-import useAutoPin from '../../../slices/navigationTabs/useAutoPin'
 
 const useMapListStyles = makeStyles((theme) => ({
     table: {
@@ -36,21 +31,22 @@ type MapListRowProps = {
 const MapListRow: FunctionComponent<MapListRowProps>= ({
     mapId
 }) => {
-    const classes = useStyles()
-    const localClasses = useMapListStyles()
-    const { url } = useRouteMatch()
-    const history = useHistory()
+    // const classes = useStyles()
+    // const localClasses = useMapListStyles()
+    const navigate = useNavigate()
 
     return <TableRow
         key={mapId}
-        onClick={() => { history.push(`${url}Edit/${mapId}/`) }}
+        onClick={() => { navigate(`Edit/${mapId}/`) }}
     >
         <TableCell>{mapId}</TableCell>
     </TableRow>
 }
 
+// eslint-disable-next-line no-empty-pattern
 type MapListProps = {}
 
+// eslint-disable-next-line no-empty-pattern
 export const MapList: FunctionComponent<MapListProps> = ({}) => {
     const classes = useStyles()
     const localClasses = useMapListStyles()
