@@ -14,7 +14,6 @@
 // Context provided:
 //
 //   - CharacterId
-//   - deactivate:  A function to deactivate the character
 //
 // NOTES:
 //
@@ -33,38 +32,28 @@
 //   TO-DO:  That.
 //
 
-import React, { useContext, useCallback, useEffect } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import React, { useContext } from 'react'
+import { useSelector } from 'react-redux'
 
-import { deactivateCharacter, registerCharacterSSM } from '../../actions/activeCharacters'
 //
 // TODO:  Rewrite activeCharacters selectors to refer to SSMs
 //
-import { getActiveCharacters, getActiveCharacterInPlayMessages } from '../../selectors/activeCharacters'
-import { getCharacters } from '../../selectors/characters'
+import { getActiveCharacters } from '../../slices/activeCharacters'
+import { getMessages } from '../../slices/messages'
+import { getCharactersInPlay } from '../../slices/ephemera'
 
 const ActiveCharacterContext = React.createContext({
     CharacterId: '',
-    deactivate: () => {}
 })
 
 export const ActiveCharacter = ({ CharacterId, children }) => {
 
-    const dispatch = useDispatch()
     const characterState = useSelector(getActiveCharacters)[CharacterId]
-    const inPlayMessages = useSelector(getActiveCharacterInPlayMessages(CharacterId))
-    const info = useSelector(getCharacters)[CharacterId]
-    //
-    // TODO:  Rewrite commands available in ActiveCharacterContext to correspond to what
-    // the SSM actually provides
-    //
-    const deactivate = useCallback(() => {
-        dispatch(deactivateCharacter(CharacterId))
-    }, [dispatch, CharacterId])
+    const inPlayMessages = useSelector(getMessages)[CharacterId]
+    const info = useSelector(getCharactersInPlay)[CharacterId]
     return (
         <ActiveCharacterContext.Provider value={{
             CharacterId,
-            deactivate,
             inPlayMessages,
             info,
             ...characterState
