@@ -1,135 +1,83 @@
-import React, { FunctionComponent, useEffect } from 'react'
+import React, { FunctionComponent } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 
 import Box from '@material-ui/core/Box'
 import TextField from '@material-ui/core/TextField'
 import Button from '@material-ui/core/Button'
-import { makeStyles } from "@material-ui/core/styles"
 
-import { characterEditByKey } from '../../selectors/characterEdit'
-import { getMyCharacterByKey } from '../../selectors/player'
-import { setValue, setDefaults, CharacterEditKeys, CharacterEditRecord } from '../../slices/characterEdit'
-import { saveCharacter, fetchCharacter } from '../../actions/UI/characterEdit'
+import { getCharacterEditDirty, getCharacterEditValues, setValue, saveCharacter } from '../../slices/UI/characterEdit'
+import { CharacterEditKeys } from '../../slices/UI/characterEdit/baseClasses'
 import useStyles from '../styles'
-
-const useCharacterEditFormStyles = makeStyles((theme) => ({
-    table: {
-        
-    }
-}))
 
 type CharacterEditFormProps = {
     characterKey: string
 }
 
 const validAssetKey = (value: string): boolean => (
-    Boolean(value.match(/^[\w\-\_\d]+$/)) && value.toLowerCase() !== 'new'
+    Boolean(value.match(/^[\w\-_\d]+$/)) && value.toLowerCase() !== 'new'
 )
-
-//
-// TODO: Create a general JS-Object to WML converter
-//
-
-//
-// TODO: Get PlayerName and include it as a player tag in the generated WML
-//
-const characterWML = (value: CharacterEditRecord['value']): string => {
-    return [
-        `<Character key="${value.assetKey}" fileName="${value.assetKey}">`,
-        `\t<Name>${value.Name}</Name>`,
-        `\t<Pronouns>${value.Pronouns}</Pronouns>`,
-        ...(value.FirstImpression ? [`\t<FirstImpression>${value.FirstImpression}</FirstImpression>`]: []),
-        ...(value.OneCoolThing ? [`\t<OneCoolThing>${value.OneCoolThing}</OneCoolThing>`]: []),
-        ...(value.Outfit ? [`\t<Outfit>${value.Outfit}</Outfit>`]: []),
-        '</Character>'
-    ].join('\n')
-}
 
 export const CharacterEditForm: FunctionComponent<CharacterEditFormProps> = ({ characterKey }) => {
     const classes = useStyles()
     const dispatch = useDispatch()
-    const localClasses = useCharacterEditFormStyles()
-    const characterMeta = useSelector(getMyCharacterByKey(characterKey))
+    // const localClasses = useCharacterEditFormStyles()
 
-    const characterEditState = useSelector(characterEditByKey(characterKey))
+    const value = useSelector(getCharacterEditValues(characterKey))
+    const dirty = useSelector(getCharacterEditDirty(characterKey))
 
-    useEffect(() => {
-        if (characterKey !== 'New' && !(characterEditState.fetched || characterEditState.fetching) && (characterMeta && characterMeta.fileName)) {
-            const dispatchPromise = dispatch(fetchCharacter(characterKey)) as unknown as Promise<Record<CharacterEditKeys, string>>
-            dispatchPromise.then((value) => {
-                    if (value) {
-                        dispatch(setDefaults({ characterKey, defaultValue: value }))
-                        const keys: CharacterEditKeys[] = [
-                            'assetKey',
-                            'Name',
-                            'Pronouns',
-                            'FirstImpression',
-                            'OneCoolThing',
-                            'Outfit'
-                        ]
-                        keys.forEach((label: CharacterEditKeys) => {
-                            if (value[label]) {
-                                dispatch(setValue({ characterKey, label, value: value[label] }))
-                            }
-                        })
-                    }
-                })
-        }
-    }, [dispatch, characterKey, characterEditState, characterMeta])
-    const { value } = characterEditState
-    const updateLabel = (label: CharacterEditKeys) => (event: { target: { value: string }}) => { dispatch(setValue({ characterKey, label, value: event.target.value })) }
+    const updateLabel = (label: CharacterEditKeys) => (event: { target: { value: string }}) => {
+        dispatch(setValue(characterKey)({ label, value: event.target.value }))
+    }
     return <Box className={classes.homeContents}>
         <TextField
             required
-            error={!validAssetKey(value.assetKey || '')}
+            error={!validAssetKey(value.assetKey)}
             id="assetKey"
             label="Asset Key"
-            value={value.assetKey || ''}
+            value={value.assetKey}
             onChange={updateLabel('assetKey')}
-            helperText="Asset key may not be 'New' and may contain only letters, numbers, - and _"
+            helperText={ !validAssetKey(value.assetKey) ? "Asset key may not be 'New' and may contain only letters, numbers, - and _" : undefined }
         />
         <TextField
             required
             id="name"
             label="Name"
-            value={value.Name || ''}
+            value={value.Name}
             onChange={updateLabel('Name')}
         />
         <TextField
             required
             id="pronouns"
             label="Pronouns"
-            value={value.Pronouns || ''}
+            value={value.Pronouns}
             onChange={updateLabel('Pronouns')}
         />
         <TextField
             id="firstImpression"
             label="First Impression"
-            value={value.FirstImpression || ''}
+            value={value.FirstImpression}
             onChange={updateLabel('FirstImpression')}
         />
         <TextField
             id="oneCoolThing"
             label="One Cool Thing"
-            value={value.OneCoolThing || ''}
+            value={value.OneCoolThing}
             onChange={updateLabel('OneCoolThing')}
         />
         <TextField
             id="outfit"
             label="Outfit"
-            value={value.Outfit || ''}
+            value={value.Outfit}
             onChange={updateLabel('Outfit')}
         />
         <Button
             variant="contained"
+            disabled={!dirty}
             onClick={() => {
                 //
                 // TODO: Refactor Redux so that the store has Typescript constraints
                 //
-                (dispatch as any)(saveCharacter(characterEditState, characterWML(value)))
-                    .then((url: string) => {
-                        alert(url)
-                    })
+                dispatch(saveCharacter(value.assetKey))
             }}
         >
                 Save
