@@ -1,11 +1,16 @@
-import React, { FunctionComponent } from 'react'
+import React, { FunctionComponent, useEffect } from 'react'
+import { useDispatch, useSelector } from 'react-redux'
+import CircularProgress from '@material-ui/core/CircularProgress'
 
 import {
     useParams
 } from "react-router-dom"
 
-import useAutoPin from '../../slices/navigationTabs/useAutoPin'
+import useAutoPin from '../../slices/UI/navigationTabs/useAutoPin'
 import CharacterEditForm from './EditForm'
+import { addItem, getStatus } from '../../slices/UI/characterEdit'
+import { heartbeat } from '../../slices/stateSeekingMachine/ssmHeartbeat'
+// import Spinner from '../Spinner'
 
 type CharacterEditProps = {}
 
@@ -15,6 +20,7 @@ type CharacterEditProps = {}
 // Add spinner for pending fetch state
 //
 
+// eslint-disable-next-line no-empty-pattern
 export const CharacterEdit: FunctionComponent<CharacterEditProps> = ({}) => {
 
     const { CharacterKey } = useParams<{ CharacterKey: string }>()
@@ -22,8 +28,19 @@ export const CharacterEdit: FunctionComponent<CharacterEditProps> = ({}) => {
         href: `/Character/Edit/${CharacterKey}`,
         label: CharacterKey === 'New' ? `New Character` : `Edit: ${CharacterKey}`
     })
+    const dispatch = useDispatch()
+    useEffect(() => {
+        if (CharacterKey && CharacterKey !== 'New') {
+            dispatch(addItem(CharacterKey))
+            dispatch(heartbeat)
+        }
+    }, [dispatch, CharacterKey])
 
-    return <CharacterEditForm characterKey={CharacterKey} />
+    const currentStatus = useSelector(getStatus(CharacterKey || 'none'))
+
+    return currentStatus === 'PARSED'
+        ? <CharacterEditForm characterKey={CharacterKey || ''} />
+        : <div style={{ height: "100%", width: "100%", display: "flex", alignItems: "center", justifyContent: "center" }}><div><CircularProgress /></div></div>
 
 }
 
