@@ -5,12 +5,13 @@ import PropTypes from "prop-types"
 import { makeStyles } from "@material-ui/core/styles"
 
 import VirtualMessageList from './VirtualMessageList'
-import { parseCommand } from '../../actions/behaviors'
+import { parseCommand } from '../../slices/lifeLine'
 import LineEntry from '../LineEntry'
 import useStyles from '../styles'
 import { useActiveCharacter } from '../ActiveCharacter'
-import useAutoPin from '../../slices/navigationTabs/useAutoPin'
-import { registerCharacterSSM } from '../../actions/activeCharacters'
+import useAutoPin from '../../slices/UI/navigationTabs/useAutoPin'
+import { addItem } from '../../slices/activeCharacters'
+import { heartbeat } from '../../slices/stateSeekingMachine/ssmHeartbeat'
 
 const useMessagePanelStyles = makeStyles((theme) => ({
     messagePanel: {
@@ -41,12 +42,13 @@ export const MessagePanel = () => {
     const { CharacterId, inPlayMessages, info: { Name = '???' } = {} } = useActiveCharacter()
     useAutoPin({ href: `/Character/${CharacterId}/Play`, label: `Play: ${Name}`})
     useEffect(() => {
-        dispatch(registerCharacterSSM({ CharacterId, defaultIntent: 'REGISTERED' }))
+        dispatch(addItem(CharacterId))
+        dispatch(heartbeat)
     }, [dispatch, CharacterId])
     const handleInput = useCallback((entry) => {
         dispatch(parseCommand(CharacterId)({ entry, raiseError: () => {} }))
         return true
-    }, [dispatch])
+    }, [dispatch, CharacterId])
     return <div className={localClasses.messagePanel}>
             <div className={localClasses.messagePanelContent}>
                 <VirtualMessageList messages={inPlayMessages} viewAsCharacterId={CharacterId} />
