@@ -4,14 +4,17 @@ import PropTypes from "prop-types"
 
 import VirtualMessageList from '../Message/VirtualMessageList'
 
-import { getMessages } from '../../selectors/messages'
+//
+// TODO: Refactor messages slice to handle threads
+//
+import { getMessages } from '../../slices/messages'
 
 export const ThreadView = ({
         ThreadId,
         viewAsCharacterId,
     }) => {
     const messages = useSelector(getMessages)
-    const threadMessages = useMemo(() => (messages.filter((message) => (message.ThreadId === ThreadId))), [messages, ThreadId])
+    const threadMessages = useMemo(() => (messages.threads.filter((message) => (message.ThreadId === ThreadId))), [messages, ThreadId])
     return <VirtualMessageList key={ThreadId} messages={threadMessages} viewAsCharacterId={viewAsCharacterId} />
 
 }
